# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ecommerce API",
        "description": "Tài liệu API cho hệ thống Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/my-shop/products": {
            "get": {
                "tags": [
                    "Shop / Products"
                ],
                "summary": "List products owned by the current shop",
                "operationId": "shopMyProducts",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Shop / Products"
                ],
                "summary": "Create a new product",
                "operationId": "shopCreateProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Public paginated product list",
                "operationId": "publicProductIndex",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Public product details",
                "operationId": "publicProductShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductDetailResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/v1/my-shop/products/{product}": {
            "put": {
                "tags": [
                    "Shop / Products"
                ],
                "summary": "Update an existing product",
                "operationId": "shopUpdateProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shop / Products"
                ],
                "summary": "Delete a product",
                "operationId": "shopDeleteProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateProductRequest": {
                "required": [
                    "name",
                    "slug",
                    "description",
                    "price",
                    "quantity",
                    "image"
                ],
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Smart Watch X1"
                    },
                    "slug": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "smart-watch-x1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest smart watch with health tracking"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1299000
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1499000,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 50
                    },
                    "sku": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "SW-X1-BLK",
                        "nullable": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "image": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "minItems": 1
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "required": [
                                "name",
                                "value",
                                "quantity"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "Color"
                                },
                                "value": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "Black"
                                },
                                "price_adjustment": {
                                    "type": "number",
                                    "format": "float",
                                    "minimum": 0,
                                    "example": 0
                                },
                                "quantity": {
                                    "type": "integer",
                                    "minimum": 0,
                                    "example": 25
                                },
                                "sku": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "SW-X1-BLK-42",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateProductRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Smart Watch X2"
                    },
                    "slug": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "smart-watch-x2"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1399000
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1599000,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 60
                    },
                    "sku": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "SW-X2-BLK",
                        "nullable": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "image": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "minItems": 1
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "required": [
                                "name",
                                "value",
                                "quantity"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "Color"
                                },
                                "value": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "Silver"
                                },
                                "price_adjustment": {
                                    "type": "number",
                                    "format": "float",
                                    "minimum": 0,
                                    "example": 50000
                                },
                                "quantity": {
                                    "type": "integer",
                                    "minimum": 0,
                                    "example": 30
                                },
                                "sku": {
                                    "type": "string",
                                    "maxLength": 100,
                                    "example": "SW-X2-SLV",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductDetailResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 99
                    },
                    "name": {
                        "type": "string",
                        "example": "Smart Watch X1"
                    },
                    "slug": {
                        "type": "string",
                        "example": "smart-watch-x1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest smart watch with health tracking"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1299000
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1499000
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "sku": {
                        "type": "string",
                        "example": "SW-X1-BLK"
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 5
                            },
                            "name": {
                                "type": "string",
                                "example": "Wearables"
                            }
                        },
                        "type": "object"
                    },
                    "shop": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "name": {
                                "type": "string",
                                "example": "Cool Gadgets"
                            },
                            "logo_url": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://example.com/logo.png"
                            },
                            "address": {
                                "type": "string",
                                "example": "123 Lê Lợi, Q1, TP.HCM"
                            }
                        },
                        "type": "object"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 7
                                },
                                "url": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://example.com/img1.jpg"
                                },
                                "order_position": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 20
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Color"
                                },
                                "value": {
                                    "type": "string",
                                    "example": "Black"
                                },
                                "price_adjustment": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 0
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 25
                                },
                                "sku": {
                                    "type": "string",
                                    "example": "SW-X1-BLK-42"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-24T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}